[PARSING] doing light parsing

+++ raytracer/str_utils.c
@@ -24,22 +24,22 @@ int my_strcmp(char *s1, char *s2)
   return *s1 == *s2;
 }
 
-int my_strncmp(char *s1, char *s2, int n)
+int my_strncmp(char *s1, char *s2, size_t n)
 {
-  int i = 0;
+  size_t i = 0;
   for (; i < n && s1[i] != '\0' && s2[i] != '\0'; ++i)
     if (s1[i] != s2[i])
       return 0;
   return 1;
 }
 
-char *my_strtok(char *s, int *index)
+char *my_strtok(char *s, size_t *index)
 {
-  int len = my_strlen_delim(s + *index); 
+  size_t len = my_strlen_delim(s + *index); 
   char *token = calloc(len + 1, sizeof (char));
   if (!token)
     return NULL;
-  for (int i = 0; i < len; ++i, *index += 1)
+  for (size_t i = 0; i < len; ++i, *index += 1)
     token[i] = s[*index];
   *index += 1;
   return token;
